+++ ubo_app/store/main/__init__.py
@@ -16,6 +16,7 @@ class SettingsCategory(StrEnum):
     NETWORK = 'Network'
     ACCESSIBILITY = 'Accessibility'
     UTILITIES = 'Utilities'
+    REMOTE = 'Remote'
     APPS = 'Apps'
 
 
@@ -23,6 +24,7 @@ SETTINGS_ICONS = {
     SettingsCategory.NETWORK: '󰛳',
     SettingsCategory.ACCESSIBILITY: '󰖤',
     SettingsCategory.UTILITIES: '󰟻',
+    SettingsCategory.REMOTE: '󰑔',
     SettingsCategory.APPS: '󰀻',
 }